fix: OrderService 移除錯誤的 (int) 轉型，修正總金額計算截斷小數（對應 Code Review 中等 #總金額計算截斷小數）

+++ src/main/java/com/trading/platform/service/OrderService.java
@@ -64,7 +64,7 @@ public class OrderService {
         order.setUser(user);
         order.setProduct(product);
         order.setQuantity(request.getQuantity());
-        order.setTotalPrice((int) product.getPrice() * request.getQuantity());
+        order.setTotalPrice(product.getPrice() * request.getQuantity());
         return orderRepository.save(order);
     }
